+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Added schedule for checks
 - Switched to Chart.js for charts
 - Better output for explain
+- Support for MySQL timeouts
 
 ## 1.2.1